add contional compliation
